+++ src/Cards/MomSaidMyTurn.cs
@@ -33,6 +33,20 @@ namespace SWIP.Cards
                     ModdingUtils.Utils.Cards.instance.AddCardToPlayer(player, randomCard, false, "", 0f, 0f, true);
                 }
             }
+
+            // Replace this card with the inert version so it can't re-grant on swap
+            var yayCard = CardRegistry.Get("Yay It's My Turn!");
+            if (yayCard != null)
+            {
+                for (int j = data.currentCards.Count - 1; j >= 0; j--)
+                {
+                    if (data.currentCards[j].cardName == "Mom Said It's My Turn")
+                    {
+                        data.currentCards[j] = yayCard;
+                        break;
+                    }
+                }
+            }
         }
 
         public override void OnRemoveCard(Player player, Gun gun, GunAmmo gunAmmo, CharacterData data, HealthHandler health, Gravity gravity, Block block, CharacterStatModifiers characterStats)